Workaround for strange autocomplete behavior in WC/react
When using the web component build of the create builder the autocomplete
behaves strange as reported in #39. I found that the selection (the user
typed text) is reset to its default value (undefined) for unknown reason.

While I dont know the exact root of the problem I found that if we make sure
`data.selection` is in sync with the quqery string then it works in Vue as
usual and also fixes the WC/react issues.

The data.selection is set in debouncedQuerySearch and not in querySearch
becuase of the debounce the user may type some more chars while the search
starts with the older query string and this causes some jittering in the
input field.