add esbuild step for some dep's and then use those
validate-iri, cuid2 and leaflet-area-select are transpiled without issue in vite
when the code is not in node_modules. However, when we distribute the package
vite won't transpile the cjs code and the component breaks. This update
introduces a step which builds these lib's into the component source
code (I know - not great!) so that we then distribute the esm versions
of these lib's when we package and publish via npm.

/facepalm