Rework profile layout handling (#73)
* revise getLayout to handle revised layout definition

* revise layout configurations

* revise profile loading order

* rework layout handling

This is a major revision to the application of profiles layouts. The
structure of the layout configuration has changed and now each property
declares its group membership or not. This change removes difficulties
associated with layouts for entities with multiple types and simplified
the layout code significantly.

* log any errors to the console also

* add missing groups

* bugfix - check needs to be computed

* when applying layout, add missing entity properties to overflow

* handle setting properties to readonly in the profile

* refactor

* don't refresh the component on crate change

* bugfix where ready is determined or not

* fix / revise profiles

* refactor - DRY

* bugfix propertyDefinition references

* we can't check the type as this is a fallback primitive

Consider the case where the property wants a URL but the data has a text
field. The text field will be loaded but there'll be a complaint that
the type doesn't match.

* render tab contents only if tab showing

* revise layouts - define as arrays

This way we can define layouts for different entity types.

* rework getLayout to return the matching layout for an entity

* put zoom inside a try catch to swallow errors

These happen when the map is not showing on the screen.

* rework layout selection

* revise timeout

* make tab reset configurable

* set default to true