Keep current tab when profile changes with compatible layout
This is similar to how this change is handled for entities and keeping the
current tab if resetTabOnEntityChange==false.

When the profile changes, previously we automatically reset the tab to
"about". In this implementation we change to "about" if
resetTabOnEntityChange==true or if the new profile doesn't have a layout
matching the currently selected one. In this case we fall back to "about".